requirements.txt
Flask==2.2.0
google-cloud-storage==2.10.0
tensorflow==2.17.1
pillow==9.3.0
opencv-python-headless==4.6.0.66
werkzeug==2.2.3
